requirements.txt
# GUI
streamlit==1.38.0
# IMG
Pillow==10.4.0
# LLM
groq==0.11.0
